indexer/pyproject.toml
[project]
name = "indexer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.9.0",
    "llama-index-core>=0.10.0",
    "llama-index-vector-stores-qdrant>=0.2.0",
    "llama-index-embeddings-openai>=0.1.0",
    "fastembed>=0.3.0",
    "tree-sitter-language-pack>=0.13.0",
    "restate-sdk>=0.9.0",
    "hypercorn>=0.17.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["indexer*"]
